Re-organize code base
- The new and old chatbot are split into 2 files for readability,
  but are still in the same module for convenience.
- Tests are now separated from code.
- Command-line and web UIs are in top-level directory.
- README and .gitignore updated
