+++ .gitignore
@@ -2,3 +2,4 @@
 flask_session
 __pycache__/
 *.log
+chatbot_logs.log.*
