+++ chatbot_cmd.py
