+++ chatbots/chatbot.py
@@ -1,0 +1,17 @@
+class Chatbot:
+    """
+    A chatbot that helps clinicians perform calculations.
+    """
+
+    def get_architecture(self) -> str:
+        """Returns a human-readable string describing the architecture of the Chatbot."""
+        pass
+
+    def invoke(self, query: str) -> str:
+        """
+        The main entry point to interact with a Chatbot.
+        Parameters:
+        - query: the input to the Chatbot
+        Returns: the response from the Chatbot
+        """
+        pass
